refactor #150: mv GithubUrlValidator into utils
